refac: add two utility functions and refactor CompressForm component
